# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Go Chat MSA API
  version: "1.0.0"
  license:
    name: 무단 사용 금지
    url: https://choosealicense.com/no-permission/
  description: |
    Go Chat MSA 프로젝트의 REST API 명세서입니다.
    - REST API 에러는 RFC 9457 (Problem Details) 형식을 따릅니다. WebSocket 업그레이드 검증 오류는 일반 HTTP 오류 본문일 수 있습니다.
    - 보안을 위해 JWT(Access Token)와 HttpOnly Cookie(Refresh Token)를 혼합하여 사용합니다.
    - 웹소켓 연결은 1회용 단기 티켓 기반으로 수행됩니다.
servers:
  - url: /api
    description: Kubernetes Ingress를 통한 API Gateway 경로
...
paths:
...
  /me/rooms:
    get:
      tags:
        - Room
      summary: 참여 중인 채팅방 목록 조회
      description: 현재 인증된 사용자가 참여 중인 채팅방 목록을 반환합니다.
      operationId: listJoinedRooms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /rooms:
    get:
      tags:
        - Room
      summary: 채팅방 검색
      description: 전체 채팅방을 대상으로 이름 기반 검색을 수행합니다.
      operationId: searchRooms
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: "검색어 (생략 시 전체 채팅방 목록 반환)"
          schema:
            type: string
        - name: limit
          in: query
          description: 페이지당 개수 (기본 20, 최대 100)
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: 검색 시작 위치
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 검색 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    post:
      tags:
        - Room
      summary: 채팅방 생성
      description: 새로운 채팅방을 만듭니다. 방을 만든 사람은 자동으로 방장이 되며 해당 방에 참여합니다.
      operationId: createRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRoomResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /rooms/{id}:
    patch:
      tags:
        - Room
      summary: 채팅방 정보 수정
      description: 방장만 호출 가능하며, 채팅방의 이름과 최대 정원을 변경합니다.
      operationId: updateRoom
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 채팅방 UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '204':
          description: 수정 성공
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      tags:
        - Room
      summary: 채팅방 삭제
      description: 방장만 호출 가능하며, 채팅방을 삭제하고 해당 방의 모든 웹소켓 연결을 강제 종료합니다.
      operationId: deleteRoom
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 채팅방 UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /rooms/{id}/members:
    get:
      tags:
        - Room
      summary: 채팅방 멤버 목록 조회
      description: 채팅방에 참여 중인 멤버 목록을 반환합니다. 요청자가 해당 방의 멤버여야 합니다.
      operationId: listRoomMembers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 채팅방 UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 멤버 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoomMemberInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /rooms/{id}/members/me:
    put:
      tags:
        - Room
      summary: 채팅방 참여
      description: 대상 채팅방에 참여합니다. 인원이 꽉 찼다면 실패합니다. (이미 참여 중인 경우에도 성공 반환)
      operationId: joinRoom
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 채팅방 UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 참여 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 정원 초과로 참여 불가
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      tags:
        - Room
      summary: 채팅방 나가기
      description: 채팅방에서 퇴장합니다.
      operationId: leaveRoom
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 채팅방 UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 퇴장 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /rooms/{id}/messages:
    get:
      tags:
        - Message
      summary: 메시지 내역 조회 및 동기화
      description: |
        채팅방의 메시지를 가져옵니다. 본인이 방에 참여한 시점 이후의 메시지만 조회됩니다.
        - 과거 내역 조회: `after_id` 없이 호출하면 가장 최근 메시지부터 id 역순으로 반환합니다.
        - 누락분 동기화: `after_id`를 포함하면 해당 id 이후를 id 정순으로 반환합니다.
      operationId: listMessages
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 채팅방 UUID
          schema:
            type: string
            format: uuid
        - name: after_id
          in: query
          description: 이 id보다 큰 메시지를 id 정순으로 반환. 빈 값은 참여 시점 이후 전체 동기화. 저장 지연 때문에 빈 응답은 누락 없음의 증거가 아님
          allowEmptyValue: true
          schema:
            anyOf:
              - type: string
                format: uuid
              - type: string
                enum: ['']
        - name: limit
          in: query
          description: 가져올 메시지 수. 생략 시 과거 조회 100개, 동기화 50개. 1000을 넘으면 1000개로 제한합니다.
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: 메시지 조회 성공
          content:
            application/json:
              schema:
                type: object
                required: [messages, has_more]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/MessageInfo'
                  has_more:
                    type: boolean
                    description: 이 응답 뒤에 더 받을 메시지가 있는지 여부
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          description: 메시지 조회를 처리할 수 없음. MongoDB 장애 중에도 WebSocket 송신은 가능할 수 있습니다.
          headers:
            Retry-After:
              description: 재시도 전 대기 시간 (초)
              schema:
                type: integer
                example: 1
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "API 요청 시 Header에 'Authorization: Bearer <Access Token>' 형식으로 포함합니다."
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: "HttpOnly, SameSite=Strict refresh token cookie입니다. 운영 환경에서는 Secure 속성이 추가됩니다."
    wsTicket:
      type: apiKey
      in: query
      name: ticket
      description: "웹소켓 핸드셰이크에 사용하는 30초 TTL 일회용 티켓입니다."
  schemas:
...
    CreateRoomRequest:
      type: object
      required: [name, capacity]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        capacity:
          type: integer
          minimum: 1
          maximum: 1000
    CreateRoomResponse:
      type: object
      properties:
        room_id:
          type: string
          format: uuid
    UpdateRoomRequest:
      type: object
      required: [name, capacity]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        capacity:
          type: integer
          minimum: 1
          maximum: 1000
    RoomListResponse:
      type: object
      properties:
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/RoomInfo'
        total_count:
          type: integer
          description: "총 검색 결과 수 (검색어를 사용했을 때만 포함됨)"
    RoomInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        manager_id:
          type: string
          format: uuid
        capacity:
          type: integer
        member_count:
          type: integer
          description: "채팅방에 참여 중인 멤버 수"
        joined_at:
          type: string
          format: date-time
          description: "방에 참여한 시간 (참여 중인 방 조회 시에만 포함)"
    RoomMemberInfo:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        joined_at:
          type: string
          format: date-time

    # Message
    MessageInfo:
      type: object
      required: [id, room_id, sender_id, content, type, timestamp]
      properties:
        id:
          type: string
          format: uuid
          description: 메시지 고유 식별자 (UUID v7). 사전순으로 정렬하며 Pod 간 인과 순서나 도착 순서를 보장하지 않음
        room_id:
          type: string
          format: uuid
        sender_id:
          type: string
          description: 작성자 user id. 시스템 메시지는 `system`
        content:
          type: string
        client_msg_id:
          type: string
          description: 클라이언트가 생성한 멱등성 키
        type:
          type: string
          description: "메시지 종류 ('chat' 또는 'system')"
        timestamp:
          type: integer
          format: int64
          description: 작성 시각 (Unix 초)

    # Common
    ProblemDetails:
      type: object
      description: 에러 발생 시 반환되는 표준 형식 (RFC 9457)
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "필수 값이 누락되었습니다."
  responses:
    BadRequest:
      description: "요청 값이 잘못됨 (예: 규칙 위반, 누락 등)"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: "인증 실패 (토큰 없음, 만료, 변조)"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: "접근 권한 없음 (예: 방장이 아님)"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: "요청한 대상을 찾을 수 없음"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    TooManyRequests:
      description: "요청 빈도 초과 (Rate Limit)"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...